Refactor code out of publish to project locator. Later we could use this class for dotnet run to reduce code duplication. Also left expansion open for sln code if we find a faster way to do it, though that way will probably be inside of MSBuild.
