+++ src/Cli/dotnet/commands/dotnet-publish/LocalizableStrings.resx
@@ -151,7 +151,4 @@ The default is to publish a framework-dependent application.</value>
   <data name="CmdNoLogo" xml:space="preserve">
     <value>Do not display the startup banner or the copyright message.</value>
   </data>
-  <data name="ProjectDeductionFailure" xml:space="preserve">
-    <value>A problem with the project or solution prevented checking properties for Release.</value>
-  </data>
 </root>
