+++ src/Cli/dotnet/commands/dotnet-publish/xlf/LocalizableStrings.tr.xlf
@@ -47,11 +47,6 @@
         <target state="translated">Yayımlamadan önce projeyi derlemeyin. --no-restore anlamına gelir.</target>
         <note />
       </trans-unit>
-      <trans-unit id="ProjectDeductionFailure">
-        <source>A problem with the project or solution prevented checking properties for Release.</source>
-        <target state="new">A problem with the project or solution prevented checking properties for Release.</target>
-        <note />
-      </trans-unit>
       <trans-unit id="RuntimeOptionDescription">
         <source>The target runtime to publish for. This is used when creating a self-contained deployment.
 The default is to publish a framework-dependent application.</source>
