+++ src/Cli/dotnet/xlf/CommonLocalizableStrings.de.xlf
@@ -120,6 +120,11 @@ export PATH="$PATH:{0}"
         <target state="translated">Für das Projekt ist bereits ein Verweis auf "{0}" vorhanden.</target>
         <note />
       </trans-unit>
+      <trans-unit id="ProjectDeductionFailure">
+        <source>A problem with the project or solution prevented checking properties for Release.</source>
+        <target state="new">A problem with the project or solution prevented checking properties for Release.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ProjectReferenceCouldNotBeFound">
         <source>Project reference `{0}` could not be found.</source>
         <target state="translated">Der Projektverweis "{0}" wurde nicht gefunden.</target>
