+++ src/Cli/dotnet/xlf/CommonLocalizableStrings.ja.xlf
@@ -120,6 +120,11 @@ export PATH="$PATH:{0}"
         <target state="translated">プロジェクトには既に `{0}` への参照が指定されています。</target>
         <note />
       </trans-unit>
+      <trans-unit id="ProjectDeductionFailure">
+        <source>A problem with the project or solution prevented checking properties for Release.</source>
+        <target state="new">A problem with the project or solution prevented checking properties for Release.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ProjectReferenceCouldNotBeFound">
         <source>Project reference `{0}` could not be found.</source>
         <target state="translated">プロジェクト参照 `{0}` は見つかりませんでした。</target>
