+++ src/Cli/dotnet/xlf/CommonLocalizableStrings.ru.xlf
@@ -120,6 +120,11 @@ export PATH="$PATH:{0}"
         <target state="translated">Проект уже содержит ссылку на "{0}".</target>
         <note />
       </trans-unit>
+      <trans-unit id="ProjectDeductionFailure">
+        <source>A problem with the project or solution prevented checking properties for Release.</source>
+        <target state="new">A problem with the project or solution prevented checking properties for Release.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ProjectReferenceCouldNotBeFound">
         <source>Project reference `{0}` could not be found.</source>
         <target state="translated">Не удалось найти ссылку на проект "{0}".</target>
